# Improve `Oj.dump` performance