@Watson1978 2021-07-25
I retrieved a benchmark result on Intel Mac.

## Env
- MacBook Pro (16-inch, 2019)
- macOS BigSur 11.5
- CPU 2.4 GHz 8cores Intel Core i9

## Before

```
Warming up --------------------------------------
             Oj.dump    22.840k i/100ms
    Oj.dump (compat)     7.228k i/100ms
     Oj.dump (rails)     7.017k i/100ms
Calculating -------------------------------------
             Oj.dump    205.030k (± 5.9%) i/s -      1.028M in   5.030663s
    Oj.dump (compat)     83.240k (± 4.5%) i/s -    419.224k in   5.047241s
     Oj.dump (rails)     71.655k (±10.1%) i/s -    357.867k in   5.077355s
```

## After
```
Warming up --------------------------------------
             Oj.dump    27.699k i/100ms
    Oj.dump (compat)     9.134k i/100ms
     Oj.dump (rails)     8.325k i/100ms
Calculating -------------------------------------
             Oj.dump    278.205k (± 3.8%) i/s -      1.413M in   5.085761s
    Oj.dump (compat)     92.628k (± 3.9%) i/s -    465.834k in   5.037398s
     Oj.dump (rails)     81.348k (± 4.6%) i/s -    407.925k in   5.025928s
```