@ohler55 2021-07-25
I made a few comments but please don't take that the wrong way. Your catch on the use of strncpy was a good one. I'd be curious if memcpy would be a bit better. Would you mind running the benchmarks to compare?

As for the other comments, I'd like to keep the constants on the ones that always return the same length but the two that use the sprintf return value are keepers so it is better to remove the non-used calculations of len earlier in the code.