@ohler55 2021-07-25
Thank you for the nice improvement. That will go nicely with the parser re-write I'm in the middle of but I'll release your improvement sooner than the parser.